DEVOPS-596-send-in-the-payload-on-label
Dearest Reviewer,

I now see this is for all events not just labeled.. I think we should
pass in the original payload and let the caller pick out what they need.
for example i know I need the pull requests html_link.

changes
pass the payload
update the label code only. not sure if this kills the rest of it?

Becker